requirements.txt
uvicorn
fastapi
python-multipart
sqlalchemy
qdrant
pydantic
python-dotenv
tree-sitter-languages
